feat: 서비스 로직 추가, 리졸버 복수 이름 단수 이름으로 변경
