+++ src/domains/comments/resolvers/addReply.js
@@ -1,10 +1,10 @@
-import {editComment} from '../services/comment.service.js';
+import {addReply} from '../services/comment.service.js';
 
 const resolver = async (parent, args, context, info) => {
   const {_id, comment, author} = args;
   const {ipv4} = context;
 
-  return editComment(_id, comment, author, ipv4);
+  return addReply(_id, comment, author, ipv4);
 };
 
 export default resolver;
