+++ src/domains/comments/resolvers/index.js
@@ -2,10 +2,10 @@ import comment from './comment.js';
 import comments from './comments.js';
 import replies from './replies.js';
 
-import addComments from './addComments.js';
+import addComment from './addComment.js';
 import editComment from './editComment.js';
 import removeComment from './removeComment.js';
-import addReplies from './addReplies.js';
+import addReply from './addReply.js';
 
 const commentResolvers = {
   Query: {
@@ -15,10 +15,10 @@ const commentResolvers = {
   },
 
   Mutation: {
-    addComments,
+    addComment,
     editComment,
     removeComment,
-    addReplies,
+    addReply,
   },
 };
