+++ src/domains/comments/resolvers/removeComment.js
@@ -1,25 +1,10 @@
-import {Comment} from '../models/index.js';
+import {removeComment} from '../services/comment.service.js';
 
-const removeComment = async (parent, args, context, info) => {
+const resolver = async (parent, args, context, info) => {
   const {_id} = args;
   const {ipv4} = context;
 
-  const filter = {
-    _id,
-    ipv4,
-    isDeleted: false,
-  };
-
-  const update = {
-    isDeleted: true,
-    deletedAt: Date.now(),
-  };
-
-  const comment = await Comment.findOneAndUpdate(filter, update, {
-    new: true,
-  });
-
-  return comment;
+  return removeComment(_id, ipv4);
 };
 
-export default removeComment;
+export default resolver;
